Move fullscreen related functionality to fullscreen_controller.js
