+++ app/frontend/startup.js
@@ -191,13 +191,6 @@ class Startup {
 
     tags_controller.initTagsUI();
     uiHelper.configureButtonStyles();
-
-    if (!platformHelper.isMac()) {
-      $('.fullscreen-button').bind('click', () => {
-        app_controller.toggleFullScreen();
-      });
-    }
-
     $(window).bind("resize", () => { uiHelper.resizeAppContainer(); });
   }
